added state to app.jsx
