+++ src/Landing.jsx
@@ -1,49 +1,51 @@
-import { Grid, Typography } from "@mui/material";
+import { Box, Grid, Typography } from "@mui/material";
 import Button from "@mui/material/Button";
 import { useNavigate } from "react-router-dom";
 
-export const Landing = () => {
+export const Landing = ({ useremail }) => {
   const navigate = useNavigate();
   return (
-    <div>
+    <Box>
       <Grid container style={{ padding: "5vw" }}>
         <Grid item xs={12} md={6} lg={6}>
-          <div style={{ marginTop: 100 }}>
+          <Box style={{ marginTop: 100 }}>
             <Typography variant={"h2"}>Coursed Admin</Typography>
             <Typography variant={"h5"}>
               A place to learn, earn and grow
             </Typography>
-            <div style={{ display: "flex", marginTop: 20 }}>
-              <div style={{ marginRight: 10 }}>
-                <Button
-                  size={"large"}
-                  variant={"contained"}
-                  onClick={() => {
-                    navigate("/signup");
-                  }}
-                >
-                  Signup
-                </Button>
-              </div>
-              <div>
-                <Button
-                  size={"large"}
-                  variant={"contained"}
-                  onClick={() => {
-                    navigate("/signin");
-                  }}
-                >
-                  Signin
-                </Button>
-              </div>
-            </div>
-          </div>
-          <div></div>
+
+            {!useremail && (
+              <Box style={{ display: "flex", marginTop: 20 }}>
+                <Box style={{ marginRight: 10 }}>
+                  <Button
+                    size={"large"}
+                    variant={"contained"}
+                    onClick={() => {
+                      navigate("/signup");
+                    }}
+                  >
+                    Signup
+                  </Button>
+                </Box>
+                <Box>
+                  <Button
+                    size={"large"}
+                    variant={"contained"}
+                    onClick={() => {
+                      navigate("/signin");
+                    }}
+                  >
+                    Signin
+                  </Button>
+                </Box>
+              </Box>
+            )}
+          </Box>
         </Grid>
         <Grid item xs={12} md={6} lg={6} style={{ marginTop: 20 }}>
           {/* <img /> */}
         </Grid>
       </Grid>
-    </div>
+    </Box>
   );
 };
